+++ overkill/overkill.py
@@ -5,14 +5,20 @@ See the :class:`ClusterCompute` for more details on distributing tasks.
 import socket
 from typing import Callable, Dict, List, Tuple, Union
 
-from overkill.utils import server_messaging_standards, utils
+from overkill.utils.server_exceptions import NoWorkersError, WorkError
+from overkill.utils.server_messaging_standards import (_DISTRIBUTE,
+                                                       _FINISHED_TASK,
+                                                       _NO_WORKERS_ERROR,
+                                                       _WORK_ERROR)
+from overkill.utils.utils import _decode_message, _encode_dict, _recv_msg, _socket_send_message
 
 
 class ClusterCompute:
     """The ClusterCompute class acts as the main interface between the Master server and the user.
     use this class to distribute an array over a cluster of computers given a function.
 
-    :param n_workers: number of workers to use to distribute (array is distributed evenly accross workers)
+    :param n_workers: number of workers to use to distribute 
+        (array is distributed evenly accross workers)
     :type n_workers: int
     :param master_address: A tuple of (ip, port) e.g. ("localhost", 5555).
         Use the .get_address() class member of the Master class to get the address
@@ -35,13 +41,13 @@ class ClusterCompute:
         """
         with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as sock:
             connection_message = {
-                "type": server_messaging_standards.DISTRIBUTE,
+                "type": _DISTRIBUTE,
                 "function": function,
                 "array": array
             }
             sock.connect(self.master_address)
-            sock.sendall(utils.encode_dict(connection_message))
-            result = utils.decode_message(sock.recv(10000))
+            _socket_send_message(_encode_dict(connection_message), sock)
+            result = _decode_message(_recv_msg(sock))
         return self.__handle_result(result)
 
     def __handle_result(self, result: Dict) -> List:
@@ -56,5 +62,9 @@ class ClusterCompute:
         :rtype: List
         """
         return_type = result.get("type")
-        if return_type == server_messaging_standards.FINISHED_TASK:
+        if return_type == _FINISHED_TASK:
             return result.get("data")
+        if return_type == _WORK_ERROR:
+            raise WorkError(result.get("error"))
+        if return_type == _NO_WORKERS_ERROR:
+            raise NoWorkersError("There are no workers connected to master")