+++ overkill/servers/worker.py
@@ -1,5 +1,7 @@
 """This module acts as the main class to create and manage a worker server"""
 
+from base64 import encode
+from inspect import trace
 import logging
 import multiprocessing
 import socketserver
@@ -7,10 +9,12 @@ import threading
 import traceback
 from typing import Dict, List, Tuple
 
-from overkill.utils.server_data_classes import masterInfo
-from overkill.utils.server_exceptions import AskTypeNotFoundError, ServerNotStartedError
-from overkill.utils.server_messaging_standards import *
-from overkill.utils.utils import *
+from overkill.utils.server_data_classes import _MasterInfo
+from overkill.utils.server_exceptions import AskTypeNotFoundError, ServerNotStartedError, WorkError
+from overkill.utils.server_messaging_standards import (_NEW_CONNECTION, _REJECT, _ACCEPT,
+                                                       _ACCEPT_WORK, _CLOSE_CONNECTION,
+                                                       _DELEGATE_WORK, _WORK_ERROR)
+from overkill.utils.utils import _recv_msg, _send_message, _encode_dict, _decode_message
 
 _master = None  # master info
 _id = None  # worker id
@@ -32,27 +36,30 @@ class _WorkerServer(socketserver.BaseRequestHandler):
         :raises askTypeNotFoundError: occurs when there is no case for an ask type
         """
         global _master, _id
-        # TODO: what happens if there's more than 1024 bytes
-        data = self.request.recv(1024)
+        data = _recv_msg(self.request)
         try:
-            ask = decode_message(data)
+            ask = _decode_message(data)
             logging.info(ask)
-            if ask["type"] == REJECT:
+            if ask["type"] == _REJECT:
                 raise Exception("Worker rejected")
-            elif ask["type"] == ACCEPT:
+            elif ask["type"] == _ACCEPT:
                 address = ask["master_address"]
-                _master = masterInfo(address)
+                _master = _MasterInfo(address)
                 logging.info(f"Master information: {_master}")
                 _id = ask["id"]
-            elif ask["type"] == DELEGATE_WORK:
+            elif ask["type"] == _DELEGATE_WORK:
                 results = self.do_work(ask)
-                send_message(encode_dict(
-                    {"type": ACCEPT_WORK, "work_id": ask["work_id"], "data": results, "order": ask["order"]}), _master.address)
+                _send_message(_encode_dict(
+                    {"type": _ACCEPT_WORK, "work_id": ask["work_id"], "data": results, "order": ask["order"]}), _master.address)
             else:
                 raise AskTypeNotFoundError(f"No such type {ask['type']}")
         except AskTypeNotFoundError as e:
             logging.info(f"No such ask exists: {e}")
             return
+        except WorkError as e:
+            logging.info(f"Encountered work error: {e}")
+            err = {"type": _WORK_ERROR, "work_id": ask["work_id"], "error": e}
+            _send_message(_encode_dict(err), _master.address)
         except Exception as e:
             logging.info(f"Could not handle request: {e}")
             logging.info(traceback.format_exc())
@@ -69,7 +76,10 @@ class _WorkerServer(socketserver.BaseRequestHandler):
         func = ask["function"]
         data = ask["array"]
 
-        results = list(map(func, data))
+        try:
+            results = list(map(func, data))
+        except Exception as e:
+            raise WorkError(f"Could not compute: {e} \n {traceback.format_exc()}")
 
         return results
 
@@ -118,9 +128,9 @@ class Worker:
         if self._server is None:
             logging.warning("No server has been started")
         else:
-            msg = encode_dict({"type": CLOSE_CONNECTION, "id": _id})
+            msg = _encode_dict({"type": _CLOSE_CONNECTION, "id": _id})
             if _master:
-                send_message(msg, _master.address)
+                _send_message(msg, _master.address)
             self._server.socket.close()
             self._server.shutdown()
             logging.info("Worker shutdown")
@@ -144,9 +154,10 @@ class Worker:
         """
         if self._server is None:
             raise ServerNotStartedError("No server has been started")
-        connection_message = {"type": NEW_CONNECTION,
+        connection_message = {"type": _NEW_CONNECTION,
                               "name": self.name, "address": self.get_address()}
         try:
-            send_message(encode_dict(connection_message), (ip, port))
+            _send_message(_encode_dict(connection_message), (ip, port))
         except ConnectionRefusedError:
-            raise ConnectionRefusedError("Please check if the master address is correct")
+            raise ConnectionRefusedError(
+                "Please check if the master address is correct")