+++ overkill/utils/server_data_classes.py
@@ -2,31 +2,34 @@
 
 from dataclasses import dataclass, field
 from threading import Event
-from typing import Any, List, Tuple
+from typing import Any, List, Tuple, Union
+
+from overkill.utils.server_exceptions import WorkError
 
 
 @dataclass
-class workerInfo:
+class _WorkerInfo:
     """Worker inforamtion class"""
     id: str
     name: str
     address: Tuple
-    active: bool = True
+    can_accept_work: bool = True
 
 
 @dataclass
-class masterInfo:
+class _MasterInfo:
     """Master information class"""
     address: Tuple
 
 
 @dataclass
-class WorkOrder:
+class _WorkOrder:
     """Class documents information about a work order from a client"""
     num_workers: int
     event: Event
     data: List = field(default_factory=lambda: [])
     progress: float = 0
+    error: Union[None, WorkError] = None
 
     def update(self, new_data: List, order: int) -> float:
         """Update the work order given new data