+++ overkill/utils/server_messaging_classes.py
@@ -2,17 +2,17 @@
 
 from dataclasses import dataclass
 from typing import Callable, List
-from overkill.utils.server_messaging_standards import DISTRIBUTE, NEW_CONNECTION
+from overkill.utils.server_messaging_standards import _DISTRIBUTE, _NEW_CONNECTION
 
 
 @dataclass(frozen=True)
 class ask_distribute:
     function: Callable
     array: List
-    type: str = DISTRIBUTE
+    type: str = _DISTRIBUTE
 
 
 @dataclass(frozen=True)
 class ask_new_connection:
     name: str
-    type: str = NEW_CONNECTION
+    type: str = _NEW_CONNECTION