+++ overkill/utils/server_messaging_standards.py
@@ -4,19 +4,18 @@ certains standards of how servers should
 interact with each other
 """
 
-# Universal encoding
-ENCODING = "utf-8"
-
 # User messaging standards:
-DISTRIBUTE = "distribute" # ask master to distribute task
+_DISTRIBUTE = "distribute"  # ask master to distribute task
 
 # Master messaging standards:
-NEW_CONNECTION = "new_connect" # recieve connection from a worker
-REJECT = "reject" # reject a worker that is trying to connect to master
-ACCEPT = "accept" # accept a worker that is trying to connect to master
-DELEGATE_WORK = "delegate_work" # delegate work to a worker
-FINISHED_TASK = "finished_task" # task from user has been completely finished
+_NEW_CONNECTION = "new_connect"  # recieve connection from a worker
+_REJECT = "reject"  # reject a worker that is trying to connect to master
+_ACCEPT = "accept"  # accept a worker that is trying to connect to master
+_DELEGATE_WORK = "delegate_work"  # delegate work to a worker
+_FINISHED_TASK = "finished_task"  # task from user has been completely finished
+_NO_WORKERS_ERROR = "no_workers_error" # error when there are no workers when the user asks for work
 
 # Worker messsaging standards:
-CLOSE_CONNECTION = "close" # close connection with master
-ACCEPT_WORK = "accept_work" # accept work from a worker
+_CLOSE_CONNECTION = "close"  # close connection with master
+_ACCEPT_WORK = "accept_work"  # accept work from a worker
+_WORK_ERROR = "work_error" # error when computing the function on the array