+++ HISTORY.rst
@@ -25,3 +25,24 @@ History
 
 * Added initial user api (i.e. ClusterCompute class and map functionality)
 
+
+0.3.1 (2022-04-18)
+------------------
+
+* Refactored utils into protected members (see issue #17)
+
+0.3.2 (2022-04-18)
+------------------
+
+* Added error handling for when the user incorrectly specifies the function or array (see issue #5)
+
+0.3.3 (2022-04-18)
+------------------
+
+* Refactored socket send and recieve protocols to handle large data
+Notes: could be optimised further but experimentation is required
+
+0.3.4 (2022-04-18)
+------------------
+* Added error when a user tries to distribute a task when there are no workers,
+user should now recieve a NoWorkersError